refactor(tests): extract inline test stubs into Go\Proxy\Generator\Stubs namespace
Move all inline test stub classes and helper functions out of test files
into dedicated files under tests/Go/Proxy/Generator/Stubs/, following PHP
standards for separating test fixtures from test logic.
