+++ tests/Go/Proxy/Generator/FunctionGeneratorTest.php
@@ -12,46 +12,17 @@ declare(strict_types=1);
 
 namespace Go\Proxy\Generator;
 
-use Exception;
 use PHPUnit\Framework\TestCase;
 use PhpParser\Node\Stmt\Function_;
 use ReflectionFunction;
 
-/**
- * Test functions for FunctionGenerator.
- */
-function funcGenHelper_simple(string $name, int $count = 0): string
-{
-    return str_repeat($name, $count);
-}
-
-function funcGenHelper_nullable(?string $x = null): ?string
-{
-    return $x;
-}
-
-function funcGenHelper_variadic(string ...$items): array
-{
-    return $items;
-}
-
-function funcGenHelper_void(): void {}
-
-function funcGenHelper_classReturn(): Exception
-{
-    return new Exception();
-}
-
-#[\Deprecated]
-function funcGenHelper_deprecated(): void {}
-
-function funcGenHelper_noAttr(): void {}
-
 class FunctionGeneratorTest extends TestCase
 {
+    private const STUBS_NS = 'Go\Proxy\Generator\Stubs';
+
     public function testFromReflectionSimple(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_simple'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_simple'));
         $output = $gen->generate();
         $this->assertStringContainsString('function funcGenHelper_simple', $output);
         $this->assertStringContainsString('string $name', $output);
@@ -61,7 +32,7 @@ class FunctionGeneratorTest extends TestCase
 
     public function testFromReflectionNullable(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_nullable'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_nullable'));
         $output = $gen->generate();
         $this->assertStringContainsString('?string $x = null', $output);
         $this->assertStringContainsString(': ?string', $output);
@@ -69,35 +40,35 @@ class FunctionGeneratorTest extends TestCase
 
     public function testFromReflectionVariadic(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_variadic'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_variadic'));
         $output = $gen->generate();
         $this->assertStringContainsString('string ...$items', $output);
     }
 
     public function testFromReflectionVoidReturn(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_void'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_void'));
         $output = $gen->generate();
         $this->assertStringContainsString(': void', $output);
     }
 
     public function testFromReflectionClassReturn(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_classReturn'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_classReturn'));
         $output = $gen->generate();
         $this->assertStringContainsString(': \Exception', $output);
     }
 
     public function testSetAndGetBody(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_simple'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_simple'));
         $gen->setBody("return 'hello';");
         $this->assertStringContainsString("return 'hello'", $gen->getBody());
     }
 
     public function testSetAndGetStmts(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_simple'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_simple'));
         $gen->setBody("return 'x';");
         $stmts = $gen->getStmts();
         $this->assertNotNull($stmts);
@@ -106,18 +77,18 @@ class FunctionGeneratorTest extends TestCase
 
     public function testSetStmtsFromArray(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_simple'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_simple'));
         $gen->setBody("return 'original';");
         $stmts = $gen->getStmts();
 
-        $gen2 = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_simple'));
+        $gen2 = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_simple'));
         $gen2->setStmts($stmts);
         $this->assertStringContainsString("return 'original'", $gen2->getBody());
     }
 
     public function testGetNode(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_simple'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_simple'));
         $node = $gen->getNode();
         $this->assertInstanceOf(Function_::class, $node);
         $this->assertSame('funcGenHelper_simple', (string) $node->name);
@@ -125,13 +96,13 @@ class FunctionGeneratorTest extends TestCase
 
     public function testGetName(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_simple'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_simple'));
         $this->assertSame('funcGenHelper_simple', $gen->getName());
     }
 
     public function testSetDocBlock(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_simple'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_simple'));
         $gen->setDocBlock(new DocBlockGenerator('My function.'));
         $output = $gen->generate();
         $this->assertStringContainsString('My function.', $output);
@@ -139,13 +110,13 @@ class FunctionGeneratorTest extends TestCase
 
     public function testGetBodyEmptyWhenNoStmts(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_simple'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_simple'));
         $this->assertSame('', $gen->getBody());
     }
 
     public function testAddParameter(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_void'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_void'));
         $param = new ParameterGenerator('extra', TypeGenerator::fromTypeString('bool'));
         $gen->addParameter($param);
         $output = $gen->generate();
@@ -154,7 +125,7 @@ class FunctionGeneratorTest extends TestCase
 
     public function testSetReturnsReference(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_simple'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_simple'));
         $gen->setReturnsReference(true);
         $output = $gen->generate();
         $this->assertStringContainsString('function &funcGenHelper_simple', $output);
@@ -163,7 +134,7 @@ class FunctionGeneratorTest extends TestCase
     public function testWideningMode(): void
     {
         $gen = FunctionGenerator::fromReflection(
-            new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_simple'),
+            new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_simple'),
             true
         );
         $output = $gen->generate();
@@ -173,7 +144,7 @@ class FunctionGeneratorTest extends TestCase
 
     public function testSetBodyEmptyString(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_simple'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_simple'));
         $gen->setBody('');
         $stmts = $gen->getStmts();
         $this->assertIsArray($stmts);
@@ -182,7 +153,7 @@ class FunctionGeneratorTest extends TestCase
 
     public function testSetReturnTypeFromTypeGenerator(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_void'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_void'));
         $typeGen = TypeGenerator::fromTypeString('int');
         $gen->setReturnType($typeGen);
         $output = $gen->generate();
@@ -200,7 +171,7 @@ class FunctionGeneratorTest extends TestCase
 
     public function testFromReflectionPreservesFunctionAttribute(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_deprecated'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_deprecated'));
         $output = $gen->generate();
         $this->assertStringContainsString('Deprecated', $output);
         $this->assertStringContainsString('#[', $output);
@@ -208,7 +179,7 @@ class FunctionGeneratorTest extends TestCase
 
     public function testFromReflectionNoAttributeWhenNone(): void
     {
-        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(__NAMESPACE__ . '\funcGenHelper_noAttr'));
+        $gen = FunctionGenerator::fromReflection(new ReflectionFunction(self::STUBS_NS . '\funcGenHelper_noAttr'));
         $output = $gen->generate();
         $this->assertStringNotContainsString('#[', $output);
     }
